requirements.txt
# GitHub Integration
pygithub>=2.1.1
gitpython>=3.1.40

# LLM Providers
openai>=1.0.0
# yandex-cloud-ml-sdk  # Uncomment if using Yandex

# CLI & Config
click>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.110.0
uvicorn>=0.27.0

# Code Quality Tools
ruff>=0.1.0
black>=23.0.0
mypy>=1.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
requests>=2.31.0
pyyaml>=6.0.0
